Switch to external OpenStack cloud provider (#7)
* Switch to external OpenStack cloud provider

The in-tree cloud providers are being transitioned to external providers
using the cloud-controller-manager plugin model.  The in-tree OpenStack
provider has already been deprecated, so this switches the charms to use
the external provider and the Cinder CSI plugin.

Depends on: https://github.com/charmed-kubernetes/charm-kubernetes-master/pull/15

* Remove node cloud tags

These were from a miscommunication and aren't actually useful, since if
you can query the node tags you already have or can already access the
cloud info.  What was actually needed is SC annotations in the K8s
cluster itself, which are now handled in cdk-addons.