Add ingress-default-ssl-certificate config options (#49)
Introduces new configuration options ingress-default-ssl-certificate
and ingress-default-ssl-key to the charm. This will allow to modify
the default ssl certificate used by the ingress controller. If either
of the configuration parameter is left as empty string, the nginx
ingress controller internally creates a default self signed certificate.
These parameters are specific to nginx ingress controller.

The user provided inputs certificate, key will be stored as k8s secret
and the secret name is passed as parameter to nginx-ingress-controller.

Closes-Bug: #1859156